Refactor project structure and add packaging
