+++ Tests/__init__.py
@@ -1,1 +1,2 @@
-import os, sys; sys.path.append(os.path.dirname(os.path.realpath(__file__)))
+"""Test suite for thePrivator."""
+
