--- Tests/coreTests.py
@@ -1,40 +1,0 @@
-import unittest
-import sys
-import os
- 
-# getting the name of the directory
-# where the this file is present.
-current = os.path.dirname(os.path.realpath(__file__))
- 
-# Getting the parent directory name
-# where the current directory is present.
-parent = os.path.dirname(current)
- 
-# adding the parent directory to 
-# the sys.path.
-sys.path.append(parent)
-from core import Core
-
-class TestCore(unittest.TestCase):
-
-    def test_start(self):
-        core = Core()
-        result = []
-        core.load_profiles()
-        for profile in core.loaded_profiles:
-            result.append(profile.name)
-        self.assertTrue(result == ['test1', 'test2'])
-
-    def test_isupper(self):
-        self.assertTrue('FOO'.isupper())
-        self.assertFalse('Foo'.isupper())
-
-    def test_split(self):
-        s = 'hello world'
-        self.assertEqual(s.split(), ['hello', 'world'])
-        # check that s.split fails when the separator is not a string
-        with self.assertRaises(TypeError):
-            s.split(2)
-
-if __name__ == '__main__':
-    unittest.main()
